Day5 Completed

+++ src/main/java/pl/pacinho/adventofcode2023/utils/DigitUtils.java
@@ -3,7 +3,7 @@ package pl.pacinho.adventofcode2023.utils;
 public class DigitUtils {
     public static boolean isDigit(String value) {
         try {
-            Integer.parseInt(value);
+            Long.parseLong(value);
             return true;
         } catch (Exception e) {
             return false;
